+++ R/shinystan-package.R
@@ -41,10 +41,11 @@
 #' 
 #' @section Saving and sharing:
 #'   
-#'   The \pkg{shinystan} package allows you to store the basic components of an
+#'   The \pkg{shinystan} package allows you to store the basic components of an 
 #'   entire project (code, posterior samples, graphs, tables, notes) in a single
-#'   object. Users can save many of the plots as ggplot2 objects for further 
-#'   customization and easy integration in reports or post-processing for 
+#'   object, a \code{\link[=as.shinystan]{shinystan object}} (sso, for short).
+#'   Users can save many of the plots as ggplot2 objects for further
+#'   customization and easy integration in reports or post-processing for
 #'   publication.
 #'   
 #'   The \code{\link{deploy_shinystan}} function lets you easily deploy your own
@@ -61,6 +62,7 @@
 #' 
 #'  Check out the demo using \code{\link{launch_shinystan_demo}} or try it with
 #'  one of your own models using \code{\link{launch_shinystan}}.
+#' 
 #'    
 #' @import ggplot2
 #' @import methods