+++ NEWS.md
@@ -1,10 +1,11 @@
 # `shinystan news`
 
 ### Version 2.1.1
-* Introduce `drop_parameters` function that removes parameters from a shinystan 
+* Introduce `drop_parameters` function for removing parameters from a shinystan 
 object (useful for very large objects when you only want to look at a subset of
 parameters)
-* Additional testing
+* Add **rstanarm** to Suggests (in the DESCRIPTION file) so `rstanarm::pp_check`
+can be called.
 
 ### Version 2.1.0
 * Compatibility with recent ggplot2 update